This PR optimizes a few array methods when dealing with small arrays (less than 16 elements). These optimizations are also done by Ruby. I tried using a bigger number (like 32): in some cases the optimized version was still faster than the non-optimized one, but in some other cases it was worse. With 16 it's always faster so it's probably a good number (and why Ruby chose it).

## `-`

Code:

```crystal
require "benchmark"
require "./old_array"

size = 16

old_a1 = OldArray.new(size) { Random::DEFAULT.hex(32) }
old_a2 = OldArray.new(size) { |i| old_a1[i] }

new_a1 = Array.new(size) { |i| old_a1[i] }
new_a2 = Array.new(size) { |i| old_a1[i] }

Benchmark.ips do |x|
  x.report("old") { old_a1 - old_a2 }
  x.report("new") { new_a1 - new_a2 }
end
```

Result:

```
old 784.20k (  1.28µs) (± 7.36%)   864B/op   4.31× slower
new   3.38M (295.74ns) (± 6.56%)  32.0B/op        fastest
```

## `|`

Code:

```crystal
size = 8

old_a1 = OldArray.new(size) { Random::DEFAULT.hex(32) }
old_a2 = OldArray.new(size) { |i| old_a1[i] }

new_a1 = Array.new(size) { |i| old_a1[i] }
new_a2 = Array.new(size) { |i| old_a1[i] }

Benchmark.ips do |x|
  x.report("old") { old_a1 | old_a2 }
  x.report("new") { new_a1 | new_a2 }
end
```

Result:

```
old   1.68M (593.85ns) (± 5.76%)  560B/op   1.92× slower
new   3.24M (308.61ns) (±10.18%)  240B/op        fastest
```

## `&`

Code:

```crystal
require "benchmark"
require "./old_array"

size = 16

old_a1 = OldArray.new(size) { Random::DEFAULT.hex(32) }
old_a2 = OldArray.new(size) { |i| old_a1[i] }

new_a1 = Array.new(size) { |i| old_a1[i] }
new_a2 = Array.new(size) { |i| old_a1[i] }

Benchmark.ips do |x|
  x.report("old") { old_a1 & old_a2 }
  x.report("new") { new_a1 & new_a2 }
end
```

Result:

```
old 784.72k (  1.27µs) (± 5.35%)  0.99kB/op   1.56× slower
new   1.23M (814.30ns) (± 5.45%)    448B/op        fastest
```

## `uniq`

Code:

```crystal
require "benchmark"
require "./old_array"

size = 16

old_a = OldArray.new(size) { Random::DEFAULT.hex }
new_a = Array.new(size) { |i| old_a[i] }

Benchmark.ips do |x|
  x.report("old") { old_a.uniq }
  x.report("new") { new_a.uniq }
end
```

Result:

```
old   1.09M (913.60ns) (± 0.97%)  0.99kB/op   1.96× slower
new   2.14M (466.89ns) (± 2.55%)    448B/op        fastest
```

## `uniq!`

Code:

```crystal
require "benchmark"
require "./old_array"

values = Array.new(16) { Random::DEFAULT.hex(64) }

old_a1 = OldArray(typeof(values[0])).new
new_a1 = Array(typeof(values[0])).new

2.times do
  values.each do |value|
    old_a1 << value
    new_a1 << value
  end
end

Benchmark.ips do |x|
  x.report("old") { old_a1.uniq! }
  x.report("new") { new_a1.uniq! }
end
```

Result:

```
old 893.02k (  1.12µs) (± 8.08%)  832B/op   3.99× slower
new   3.56M (280.82ns) (± 2.64%)  0.0B/op        fastest
```