@Sija 2019-08-17
@asterite Is it my code or did this PR introduce a regression?

See:
- [failed job on CI](https://travis-ci.org/crystal-money/money/jobs/572971921) (same spec passes on crystal `latest`)
- [relevant spec lines](https://github.com/crystal-money/money/blob/e108d187d196c682942e12fe89534a627d85e255/spec/money_spec.cr#L252-L255)
- [def `<=>`](https://github.com/crystal-money/money/blob/e108d187d196c682942e12fe89534a627d85e255/src/money/money.cr#L86-L91)
- added in this PR [`Array#includes?` call](https://github.com/crystal-lang/crystal/pull/8048/files#diff-a375a8eca3c8195929f07569b36c7977R216)