@asterite 2019-08-17
@Sija I don't know, maybe. Do you have a case where it is broken?