@asterite 2019-08-17
I mean, show me an array method that works in a wrong way and we can fix it. Otherwise I'm a bit lost regarding what to fix.