@asterite 2019-08-17
@Sija try doing this:

```crystal
Money.new(1_00, "EUR") == Money.new(1_00, "USD")
```

It fails with the same error. `Array#&` changed a bit and in the optimization it might use `==` or a hash lookup. It seems the problem is on your side.