@Sija 2019-08-17
> It fails with the same error. `Array#&` changed a bit and in the optimization it might use `==` or a hash lookup. It seems the problem is on your side.

That's the thing, before it was using a hash lookup, whereas now it uses`<=>` through `Array#includes?` call.

> @Sija try doing this:
> 
> ```crystal
> Money.new(1_00, "EUR") == Money.new(1_00, "USD")
> ```

I know it doesn't work, but that's how it's supposed to be. Hash lookup on the other hand is suppose to work based on different semantics (comparing the attributes).