@asterite 2019-08-17
No, hash lookup uses first the hash method and then the == method. It's just that before this change a hash was built from the second array and entries were removed based on the first array so it never needed to compare two different currencies. But swap tee order of EUR and USD in that spec and I'm sure it will fail before and after this PR.

In every object you must guarantee that if two objects are equal by == then their hash code must be the same (but the reverse is not true).