@Sija 2019-08-18
> But swap the order of EUR and USD in that spec and I'm sure it will fail before and after this PR.

It works in every combination, just tried.