@asterite 2019-08-18
And with that I mean, there's no problem on our side. I can't understand why you are relying on `hash` for that to work. As I said, if two objects have the same hash it doesn't mean they are equal. If for some reason the hash seed (which changes from program to program) makes it so that `"USD".hash == "EUR".hash` then that spec will start failing (depending on the order of the elements in the array).